+++ shell/platform/android/io/flutter/plugin/common/MethodCodec.java
@@ -7,10 +7,7 @@ package io.flutter.plugin.common;
 import java.nio.ByteBuffer;
 
 /**
- * A message encoding/decoding mechanism with support for method calls and enveloped replies.
- *
- * Messages are semi-structured values whose static Java type is Object. Concrete codecs support
- * different types of messages.
+ * A codec for method calls and enveloped results.
  *
  * Method calls are encoded as binary messages with enough structure that the codec can
  * extract a method name String and an arguments Object. These data items are used to populate a
@@ -18,13 +15,13 @@ import java.nio.ByteBuffer;
  *
  * All operations throw {@link IllegalArgumentException}, if conversion fails.
  */
-public interface MethodCodec extends MessageCodec<Object> {
+public interface MethodCodec {
     /**
      * Decodes a message call from binary.
      *
-     * @param methodCall the binary encoding of the method call as a {@link ByteBuffer}, not null.
+     * @param methodCall the binary encoding of the method call as a {@link ByteBuffer}.
      * @return a {@link MethodCall} representation of the bytes between the given buffer's current
-     * position and its limit, not null.
+     * position and its limit.
      */
     MethodCall decodeMethodCall(ByteBuffer methodCall);
 
@@ -38,10 +35,10 @@ public interface MethodCodec extends MessageCodec<Object> {
     ByteBuffer encodeSuccessEnvelope(Object result);
 
     /**
-     * Encodes an error reply into a binary envelope message.
+     * Encodes an error result into a binary envelope message.
      *
-     * @param errorCode An error code String, not null.
-     * @param errorMessage An error message String, not null.
+     * @param errorCode An error code String.
+     * @param errorMessage An error message String, possibly null.
      * @param errorDetails Error details, possibly null.
      * @return a ByteBuffer containing the encoding between position 0 and
      * the current position.
